break up global store, organize files
