+++ website/app/index.tsx
@@ -4,36 +4,16 @@ import {
   Heading,
   Text,
   ChakraProps,
-  useColorMode,
   IconButton,
   VisuallyHidden,
   Flex,
 } from '@chakra-ui/react';
-import create from 'zustand';
-import { IconSearch } from './icon-search';
-import { LogoGithubIcon, LogoNpmIcon, MoonIcon, SunnyIcon } from '../../src';
-import { ChakraLogo } from './chakra-logo';
-import { IonIconLogo } from './ionicons-logo';
-import { SelectIconType } from './select-icon-type';
-import { SelectedIconType } from './icon-data-2';
-
-type AppStore = {
-  selectedIconType: SelectedIconType;
-  setSelectedIconType: (selected: SelectedIconType) => void;
-};
-
-export const useAppStore = create<AppStore>(set => ({
-  selectedIconType: 'filled',
-  setSelectedIconType: selected =>
-    set(state => ({
-      ...state,
-      selectedIconType: selected,
-    })),
-}));
+import { IconSearchSection } from './search/icon-search-section';
+import { LogoGithubIcon, LogoNpmIcon } from '../../src';
+import { ChakraLogo } from './parts/chakra-logo';
+import { IonIconLogo } from './parts/ionicons-logo';
 
 export function App() {
-  const { colorMode, toggleColorMode } = useColorMode();
-
   const fullWidthCenter: ChakraProps = {
     width: '100%',
     textAlign: 'center',
@@ -56,19 +36,7 @@ export function App() {
         <Text {...fullWidthCenter} color="gray.500" mb={[5, 7]}>
           Chakra compatible Ion Icons. Click on an icon to see usage.
         </Text>
-        <IconSearch
-          onSelect={icon => console.log(icon)}
-          options={
-            <>
-              <SelectIconType />
-              <IconButton
-                onClick={toggleColorMode}
-                aria-label="change color mode"
-                icon={colorMode === 'light' ? <MoonIcon /> : <SunnyIcon />}
-              />
-            </>
-          }
-        />
+        <IconSearchSection />
       </Box>
       <Flex
         position="absolute"
